fix: unable to resolve stream module: downgrade htmlparser2 to v3.10.1
fix #244

The commit 6ef90ee4bbb10d6894bb20713188c2bf4d5bc90c upgraded the
htmlparser2 dependency to version 4.0.0. This major release introduced a
hard dependency on the stream node module, which is not available in
react native runtime. However, it appears that when bundling the
javascript source for production, metro will sometimes strip the code
responsible for importing the stream module, and in other instances it
will keep it, triggering an error and failing.