refactor: remove misleading sub and sup styles
textAlignVertical doesn't work relatevely to line, but to containing
block. Those styles gave the false impression that sub and sup where
fully supported, but that will not be the case until React Native has
official support for this.

Vote for the feature here: https://react-native.canny.io/feature-requests/p/support-superscript-and-subscript

Related #259, #76