feat: new renderers API
The new renderer API is based on the transient render engine available
at @native-html/transient-render-engine.

This API changes the nature of renderers, which are now React component
instead of render functions taking 4 arguments. Most notably:

- The first argument has been replaced with the "tnode" prop;
- The second argument has been replaced with "children" props;
- The third argument has been replaced with "style" prop;
- The fourth argument is available with "useSharedProps" hook.

In addition, the "TDefaultRenderer" prop is a component to render the
passed tnode. Consumers can also take advantage of the TChildrenRenderer
and TNodeChildrenRenderer exported components to customize the rendering
of the component, while preserving the rendering logic of this library.

Moreover, renderers must declare a "model" static field to specify the
HTMLElementModel corresponding to the tag to render.