feat: new `markers` prop for custom renderers
Markers form an abstraction in which one node provides semantic
information to itself and all its descendants by the mean of its name or
attributes. For example, `ins` elements, which stand for "insertion" of
content in the context of an edit will provide the { edits: 'ins' }
marker to all its descendants.

Custom renderers can use markers to change their layout and convey their
semantic meaning. Markers can be derived from attributes, such as `lang`
and `dir` attributes, or tag names, such as `a`, `ins`, `del`...