feat: new dev profiler to get feedback on short updates
BREAKING CHANGE: The `triggerTREInvalidationPropNames` has been
discontinued. The idea for this prop originated in the premise that
many beginners would disregard the issue of passing literal props triggering
many re-renders. But I realized it mostly frustrated expectations of
newcomers that this library honors React components contract. So I have
finally decided to discard the prop, and instead add a lightweight
profiler (in dev mode only, of course), which warns the consumer of
re-renders happening in a short period of time. You are now responsible
for the invalidation of the TRenderEngine, so make sure you memoize the
props if the controlling components is expected to re-render often.