refactor: handle `key` in `TChildrenRenderer`
Keys don't need to be handled deeper in the tree, only when returning
arrays of React nodes. This was a legacy of v5, where custom renderers
were functions. Now that any custom renderer is wrapped in a
`TNodeRenderer` which has its own key, there is no need to use this key
in custom renderers.